Fold folderfetcher.py into create_reports.py Add version argument
